+++ Sources/CatalogAPIClient/Client/CatalogAPIClient+Person.swift
@@ -8,19 +8,19 @@ import Tracing
 extension CatalogAPIClient {
 
   public func fetchPersons() async throws -> JSONAPIDocument<PersonAttributes> {
-    try await withSpan("fetchPersons") { _ in
+    try await withSpan("fetch-persons") { _ in
       try await fetch(path: "/persons")
     }
   }
 
   public func fetchPerson(id: String) async throws -> JSONAPISingleDocument<PersonAttributes> {
-    try await withSpan("fetchPerson") { _ in
+    try await withSpan("fetch-person") { _ in
       try await fetch(path: "/persons/\(id)")
     }
   }
 
   public func fetchPersonVolumes(id: String) async throws -> JSONAPIDocument<VolumeAttributes> {
-    try await withSpan("fetchPersonVolumes") { _ in
+    try await withSpan("fetch-person-volumes") { _ in
       try await fetch(path: "/persons/\(id)/volumes")
     }
   }