+++ Sources/CatalogAPIClient/Client/CatalogAPIClient+Publisher.swift
@@ -8,7 +8,7 @@ import Tracing
 extension CatalogAPIClient {
 
   public func fetchPublishers() async throws -> JSONAPIDocument<PublisherAttributes> {
-    try await withSpan("fetchPublishers") { _ in
+    try await withSpan("fetch-publishers") { _ in
       try await fetch(path: "/publishers")
     }
   }
@@ -16,13 +16,13 @@ extension CatalogAPIClient {
   public func fetchPublisher(id: String) async throws -> JSONAPISingleDocument<
     PublisherAttributes
   > {
-    try await withSpan("fetchPublisher") { _ in
+    try await withSpan("fetch-publisher") { _ in
       try await fetch(path: "/publishers/\(id)")
     }
   }
 
   public func fetchPublisherVolumes(id: String) async throws -> JSONAPIDocument<VolumeAttributes> {
-    try await withSpan("fetchPublisherVolumes") { _ in
+    try await withSpan("fetch-publisher-volumes") { _ in
       try await fetch(path: "/publishers/\(id)/volumes")
     }
   }