+++ Sources/CatalogAPIClient/Client/CatalogAPIClient+Studio.swift
@@ -8,19 +8,19 @@ import Tracing
 extension CatalogAPIClient {
 
   public func fetchStudios() async throws -> JSONAPIDocument<StudioAttributes> {
-    try await withSpan("fetchStudios") { _ in
+    try await withSpan("fetch-studios") { _ in
       try await fetch(path: "/studios")
     }
   }
 
   public func fetchStudio(id: String) async throws -> JSONAPISingleDocument<StudioAttributes> {
-    try await withSpan("fetchStudio") { _ in
+    try await withSpan("fetch-studio") { _ in
       try await fetch(path: "/studios/\(id)")
     }
   }
 
   public func fetchStudioVolumes(id: String) async throws -> JSONAPIDocument<VolumeAttributes> {
-    try await withSpan("fetchStudioVolumes") { _ in
+    try await withSpan("fetch-studio-volumes") { _ in
       try await fetch(path: "/studios/\(id)/volumes")
     }
   }